Mon, Sep 18, 2023, 8:39 AM +01:00

+++ week9/birthdays/app.py
@@ -24,14 +24,16 @@ def after_request(response):
 def index():
     if request.method == "POST":
         #load data from form
-        name = request.form.get("name")
-        day = request.form.get("day")
-        month = request.form.get("month")
-        test = request.form.get("id")
-        #remember submitted birthday
-        db.execute("INSERT INTO birthdays (name, day, month) VALUES(?, ?, ?)", name, day, month)
-        #delete selected record
-        db.execute("DELETE FROM birthdays WHERE id='?'", test)
+        if request.form.get("name") /= NULL:
+            name = request.form.get("name")
+            day = request.form.get("day")
+            month = request.form.get("month")
+            #remember submitted birthday
+            db.execute("INSERT INTO birthdays (name, day, month) VALUES(?, ?, ?)", name, day, month)
+        if request.form.get("id") /= NULL:
+            test = request.form.get("id")
+            #delete selected record
+            db.execute("DELETE FROM birthdays WHERE id='?'", test)
         return redirect("/")
 
     else:
